requirements.txt
absl-py==2.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
azure-functions==1.22.1
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.1
bing-image-downloader==1.1.2
bleach==6.1.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
celery==5.5.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
comm==0.2.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz#sha256=14a2f31bc476af87019819ea8c9948fabdfd473a442edd6b1cba62bf0c2c0f55
executing==2.0.1
ExifRead==3.0.0
fastapi==0.115.11
fastjsonschema==2.19.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
flickrapi==2.4.0
fonttools==4.53.0
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
geosphere==0.0.1
google-pasta==0.2.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
idna==3.6
igraph==0.11.5
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
keras==3.8.0
kiwisolver==1.4.5
kombu==5.5.2
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mizani==0.11.4
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
overrides==7.6.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
patsy==0.5.6
pillow==10.3.0
platformdirs==4.1.0
plotly==5.22.0
plotnine==0.13.6
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==5.29.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.17.1
scikit-learn==1.6.1
scipy==1.13.1
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.0
soupsieve==2.5
sp==1.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spicy==0.16.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.0
statsmodels==0.14.2
sympy==1.13.1
tenacity==8.4.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.0
text-unidecode==1.3
textblob==0.19.0
texttable==1.7.0
thinc==8.3.4
threadpoolctl==3.6.0
tinycss2==1.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
typer==0.15.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.1.0
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
